+++ src/main/java/com/ironhack/MidtermProject/dao/accounts/StudentChecking.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Optional;
 
@@ -17,7 +18,7 @@ import java.util.Optional;
 @DiscriminatorValue("student")
 public class StudentChecking extends Account {
 
-    public StudentChecking(double balance, String primaryOwner, String secondaryOwner, AccountHolder accountHolder) {
+    public StudentChecking(BigDecimal balance, String primaryOwner, String secondaryOwner, AccountHolder accountHolder) {
         super(balance, primaryOwner, secondaryOwner, accountHolder);
     }
 }
